pyproject.toml
[tool.poetry]
name = "signal-editor"
version = "0.2.0"
description = ""
authors = ["Frederik Krämer <frdrkkraemer@gmail.com>"]
readme = "README.md"
packages = [{ include = "signal_editor", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pyqtgraph = { git = "https://github.com/pyqtgraph/pyqtgraph.git", rev = "master" }
pyopengl-accelerate = "^3.1.7"
pyarrow = "^14.0.1"
wfdb = "^4.1.2"
polars = "^0.20.1"
neurokit2 = "^0.2.7"
pyopengl = "^3.1.7"
mne = "^1.6.0"
pywavelets = "^1.5.0"
qdarkstyle = "^3.2.3"
pyside6 = "^6.6.0"
numpy = "^1.26.2"
scipy = "^1.11.4"
xlsxwriter = "^3.1.9"
h5py = "^3.10.0"
pyinstaller = "^6.3.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.18.1"
sourcery = "^1.14.0"
better-exceptions = "^0.3.3"
loguru = "^0.7.2"
pytest = "^7.4.3"
mypy = "^1.7.1"
ruff = "^0.1.6"
matplotlib = "^3.8.2"
hdf5view = "^0.0.7"
black = "^23.12.1"
pygments = "^2.17.2"
rich = "^13.7.0"
devtools = "^0.12.2"
qtconsole = "^5.5.1"
pdir2 = "^1.1.0"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
plugins = [
    "numpy.typing.mypy_plugin",
]

[tool.ruff]
target-version = "py312"
select = ["E", "F", "B"]
show-fixes = true
show-source = true
line-length = 100

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.lint]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [
    "--import-mode=importlib",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
pythonVersion = "3.12"
reportMissingTypeStubs = false
reportUnknownMemberType = false
